requirements.txt
absl-py==2.3.1
accelerate==1.0.1
addict==2.4.0
affine==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.16.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.45.5
bleach==6.1.0
blinker==1.8.2
boto3==1.37.38
botocore==1.37.38
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
clean-fid==0.1.35
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==3.1.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.30.1
dill==0.3.8
distlib==0.4.0
docker==7.1.0
docker-pycreds==0.4.0
einops==0.7.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
filterpy==1.4.5
Flask==3.0.3
fonttools==4.55.6
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
future==1.0.0
geojson==3.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
GPUtil==1.4.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.70.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.33.4
icecream==2.1.8
identify==2.6.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
lightning_sdk==2025.7.22
llvmlite==0.41.1
lmdb==1.7.3
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
mistune==3.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.3
nodeenv==1.9.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
open3d==0.19.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
OpenEXR==3.3.2
orjson==3.10.15
overrides==7.7.0
packaging==25.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.9.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
plyfile==1.0.3
plyflatten==0.2.0
pre-commit==3.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyiqa==0.1.14.1
PyJWT==2.9.0
pyparsing==3.1.4
pyproj==3.5.0
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.3.11
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpcm==1.4.10
rpds-py==0.20.1
rsa==4.9.1
ruff==0.9.10
s3transfer==0.11.5
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
shapely==2.0.2
shellingham==1.5.4
simple-term-menu==1.6.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.6
srtm4==1.2.5
stack-data==0.6.3
starlette==0.44.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.19
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.12.0
tornado==6.4.2
torchmetrics==1.2.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trimesh==4.6.0
triton==3.0.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.2.3
utm==0.8.1
uvicorn==0.33.0
virtualenv==20.34.0
wandb==0.19.4
wcwidth==0.2.13
webcolors==24.8.0
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.6
wget==3.2
widgetsnbextension==4.0.13
xxhash==3.5.0
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
